Combinations generator

This function generates all the combinations of selecting k items from n items. The results are in lexicographical order.

combinations(n, k, x = NULL, freq = NULL, replace = FALSE,
  type = "r")

Arguments

n
an integer. If freq is not NULL, n would be set as sum(freq), otherwise if x is not NULL, n would be set as length(x).

k
an integer

x
an optional vector indicating item labels

freq
an integer vector of item repeat frequencies

replace
an logical to draw items with replacement

type
if "r", "c" or "l" is specified, the returned value would be a "row-major" matrix, a "column-major" matrix or a list respectively

See also

Examples

# choose 2 from 4
combinations(4, 2)

#>      [,1] [,2]
#> [1,]    1    2
#> [2,]    1    3
#> [3,]    1    4
#> [4,]    2    3
#> [5,]    2    4
#> [6,]    3    4

combinations(x = LETTERS[1:3], k = 2)

#>      [,1] [,2]
#> [1,] "A"  "B" 
#> [2,] "A"  "C" 
#> [3,] "B"  "C" 

# multiset with frequencies c(2, 3)
combinations(freq = c(2, 3), k = 3)

#>      [,1] [,2] [,3]
#> [1,]    1    1    2
#> [2,]    1    2    2
#> [3,]    2    2    2

# with replacement
combinations(4, 2, replace = TRUE)

#>       [,1] [,2]
#>  [1,]    1    1
#>  [2,]    1    2
#>  [3,]    1    3
#>  [4,]    1    4
#>  [5,]    2    2
#>  [6,]    2    3
#>  [7,]    2    4
#>  [8,]    3    3
#>  [9,]    3    4
#> [10,]    4    4

# column major
combinations(4, 2, type = "c")

#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1    1    1    2    2    3
#> [2,]    2    3    4    3    4    4

# list output
combinations(4, 2, type = "l")

#> [[1]]
#> [1] 1 2
#> 
#> [[2]]
#> [1] 1 3
#> 
#> [[3]]
#> [1] 1 4
#> 
#> [[4]]
#> [1] 2 3
#> 
#> [[5]]
#> [1] 2 4
#> 
#> [[6]]
#> [1] 3 4
#> 

# zero sized combinations
dim(combinations(5, 0))

#> [1] 1 0

dim(combinations(5, 6))

#> [1] 0 6

dim(combinations(0, 0))

#> [1] 1 0

dim(combinations(0, 1))

#> [1] 0 1