Compositions generator

This function generates the compositions of an non-negative interger n into k parts or parts of any sizes. The results are in lexicographical or reversed lexicographical order.

compositions(n, k = NULL, descending = FALSE, layout = NULL,
  nitem = -1L, skip = NULL, index = NULL, nsample = NULL,
  drop = NULL)

Arguments

n
an non-negative integer to be partitioned

k
number of parts

descending
an logical to use reversed lexicographical order

layout
if "row", "column" or "list" is specified, the returned value would be a "row-major" matrix, a "column-major" matrix or a list respectively

nitem
number of compositions required, usually used with skip

skip
the number of compositions skipped

index
a vector of indices of the desired compositions

nsample
sampling random compositions

drop
vectorize a matrix or unlist a list

See also

Examples

# all compositions of 6
compositions(6)

#>       [,1] [,2] [,3] [,4] [,5] [,6]
#>  [1,]    1    1    1    1    1    1
#>  [2,]    1    1    1    1    2    0
#>  [3,]    1    1    1    2    1    0
#>  [4,]    1    1    1    3    0    0
#>  [5,]    1    1    2    1    1    0
#>  [6,]    1    1    2    2    0    0
#>  [7,]    1    1    3    1    0    0
#>  [8,]    1    1    4    0    0    0
#>  [9,]    1    2    1    1    1    0
#> [10,]    1    2    1    2    0    0
#> [11,]    1    2    2    1    0    0
#> [12,]    1    2    3    0    0    0
#> [13,]    1    3    1    1    0    0
#> [14,]    1    3    2    0    0    0
#> [15,]    1    4    1    0    0    0
#> [16,]    1    5    0    0    0    0
#> [17,]    2    1    1    1    1    0
#> [18,]    2    1    1    2    0    0
#> [19,]    2    1    2    1    0    0
#> [20,]    2    1    3    0    0    0
#> [21,]    2    2    1    1    0    0
#> [22,]    2    2    2    0    0    0
#> [23,]    2    3    1    0    0    0
#> [24,]    2    4    0    0    0    0
#> [25,]    3    1    1    1    0    0
#> [26,]    3    1    2    0    0    0
#> [27,]    3    2    1    0    0    0
#> [28,]    3    3    0    0    0    0
#> [29,]    4    1    1    0    0    0
#> [30,]    4    2    0    0    0    0
#> [31,]    5    1    0    0    0    0
#> [32,]    6    0    0    0    0    0

# reversed lexicographical order
compositions(6, descending = TRUE)

#>       [,1] [,2] [,3] [,4] [,5] [,6]
#>  [1,]    6    0    0    0    0    0
#>  [2,]    5    1    0    0    0    0
#>  [3,]    4    2    0    0    0    0
#>  [4,]    4    1    1    0    0    0
#>  [5,]    3    3    0    0    0    0
#>  [6,]    3    2    1    0    0    0
#>  [7,]    3    1    2    0    0    0
#>  [8,]    3    1    1    1    0    0
#>  [9,]    2    4    0    0    0    0
#> [10,]    2    3    1    0    0    0
#> [11,]    2    2    2    0    0    0
#> [12,]    2    2    1    1    0    0
#> [13,]    2    1    3    0    0    0
#> [14,]    2    1    2    1    0    0
#> [15,]    2    1    1    2    0    0
#> [16,]    2    1    1    1    1    0
#> [17,]    1    5    0    0    0    0
#> [18,]    1    4    1    0    0    0
#> [19,]    1    3    2    0    0    0
#> [20,]    1    3    1    1    0    0
#> [21,]    1    2    3    0    0    0
#> [22,]    1    2    2    1    0    0
#> [23,]    1    2    1    2    0    0
#> [24,]    1    2    1    1    1    0
#> [25,]    1    1    4    0    0    0
#> [26,]    1    1    3    1    0    0
#> [27,]    1    1    2    2    0    0
#> [28,]    1    1    2    1    1    0
#> [29,]    1    1    1    3    0    0
#> [30,]    1    1    1    2    1    0
#> [31,]    1    1    1    1    2    0
#> [32,]    1    1    1    1    1    1

# fixed number of parts
compositions(10, 5)

#>        [,1] [,2] [,3] [,4] [,5]
#>   [1,]    1    1    1    1    6
#>   [2,]    1    1    1    2    5
#>   [3,]    1    1    1    3    4
#>   [4,]    1    1    1    4    3
#>   [5,]    1    1    1    5    2
#>   [6,]    1    1    1    6    1
#>   [7,]    1    1    2    1    5
#>   [8,]    1    1    2    2    4
#>   [9,]    1    1    2    3    3
#>  [10,]    1    1    2    4    2
#>  [11,]    1    1    2    5    1
#>  [12,]    1    1    3    1    4
#>  [13,]    1    1    3    2    3
#>  [14,]    1    1    3    3    2
#>  [15,]    1    1    3    4    1
#>  [16,]    1    1    4    1    3
#>  [17,]    1    1    4    2    2
#>  [18,]    1    1    4    3    1
#>  [19,]    1    1    5    1    2
#>  [20,]    1    1    5    2    1
#>  [21,]    1    1    6    1    1
#>  [22,]    1    2    1    1    5
#>  [23,]    1    2    1    2    4
#>  [24,]    1    2    1    3    3
#>  [25,]    1    2    1    4    2
#>  [26,]    1    2    1    5    1
#>  [27,]    1    2    2    1    4
#>  [28,]    1    2    2    2    3
#>  [29,]    1    2    2    3    2
#>  [30,]    1    2    2    4    1
#>  [31,]    1    2    3    1    3
#>  [32,]    1    2    3    2    2
#>  [33,]    1    2    3    3    1
#>  [34,]    1    2    4    1    2
#>  [35,]    1    2    4    2    1
#>  [36,]    1    2    5    1    1
#>  [37,]    1    3    1    1    4
#>  [38,]    1    3    1    2    3
#>  [39,]    1    3    1    3    2
#>  [40,]    1    3    1    4    1
#>  [41,]    1    3    2    1    3
#>  [42,]    1    3    2    2    2
#>  [43,]    1    3    2    3    1
#>  [44,]    1    3    3    1    2
#>  [45,]    1    3    3    2    1
#>  [46,]    1    3    4    1    1
#>  [47,]    1    4    1    1    3
#>  [48,]    1    4    1    2    2
#>  [49,]    1    4    1    3    1
#>  [50,]    1    4    2    1    2
#>  [51,]    1    4    2    2    1
#>  [52,]    1    4    3    1    1
#>  [53,]    1    5    1    1    2
#>  [54,]    1    5    1    2    1
#>  [55,]    1    5    2    1    1
#>  [56,]    1    6    1    1    1
#>  [57,]    2    1    1    1    5
#>  [58,]    2    1    1    2    4
#>  [59,]    2    1    1    3    3
#>  [60,]    2    1    1    4    2
#>  [61,]    2    1    1    5    1
#>  [62,]    2    1    2    1    4
#>  [63,]    2    1    2    2    3
#>  [64,]    2    1    2    3    2
#>  [65,]    2    1    2    4    1
#>  [66,]    2    1    3    1    3
#>  [67,]    2    1    3    2    2
#>  [68,]    2    1    3    3    1
#>  [69,]    2    1    4    1    2
#>  [70,]    2    1    4    2    1
#>  [71,]    2    1    5    1    1
#>  [72,]    2    2    1    1    4
#>  [73,]    2    2    1    2    3
#>  [74,]    2    2    1    3    2
#>  [75,]    2    2    1    4    1
#>  [76,]    2    2    2    1    3
#>  [77,]    2    2    2    2    2
#>  [78,]    2    2    2    3    1
#>  [79,]    2    2    3    1    2
#>  [80,]    2    2    3    2    1
#>  [81,]    2    2    4    1    1
#>  [82,]    2    3    1    1    3
#>  [83,]    2    3    1    2    2
#>  [84,]    2    3    1    3    1
#>  [85,]    2    3    2    1    2
#>  [86,]    2    3    2    2    1
#>  [87,]    2    3    3    1    1
#>  [88,]    2    4    1    1    2
#>  [89,]    2    4    1    2    1
#>  [90,]    2    4    2    1    1
#>  [91,]    2    5    1    1    1
#>  [92,]    3    1    1    1    4
#>  [93,]    3    1    1    2    3
#>  [94,]    3    1    1    3    2
#>  [95,]    3    1    1    4    1
#>  [96,]    3    1    2    1    3
#>  [97,]    3    1    2    2    2
#>  [98,]    3    1    2    3    1
#>  [99,]    3    1    3    1    2
#> [100,]    3    1    3    2    1
#> [101,]    3    1    4    1    1
#> [102,]    3    2    1    1    3
#> [103,]    3    2    1    2    2
#> [104,]    3    2    1    3    1
#> [105,]    3    2    2    1    2
#> [106,]    3    2    2    2    1
#> [107,]    3    2    3    1    1
#> [108,]    3    3    1    1    2
#> [109,]    3    3    1    2    1
#> [110,]    3    3    2    1    1
#> [111,]    3    4    1    1    1
#> [112,]    4    1    1    1    3
#> [113,]    4    1    1    2    2
#> [114,]    4    1    1    3    1
#> [115,]    4    1    2    1    2
#> [116,]    4    1    2    2    1
#> [117,]    4    1    3    1    1
#> [118,]    4    2    1    1    2
#> [119,]    4    2    1    2    1
#> [120,]    4    2    2    1    1
#> [121,]    4    3    1    1    1
#> [122,]    5    1    1    1    2
#> [123,]    5    1    1    2    1
#> [124,]    5    1    2    1    1
#> [125,]    5    2    1    1    1
#> [126,]    6    1    1    1    1

# reversed lexicographical order
compositions(10, 5, descending = TRUE)

#>        [,1] [,2] [,3] [,4] [,5]
#>   [1,]    6    1    1    1    1
#>   [2,]    5    2    1    1    1
#>   [3,]    5    1    2    1    1
#>   [4,]    5    1    1    2    1
#>   [5,]    5    1    1    1    2
#>   [6,]    4    3    1    1    1
#>   [7,]    4    2    2    1    1
#>   [8,]    4    2    1    2    1
#>   [9,]    4    2    1    1    2
#>  [10,]    4    1    3    1    1
#>  [11,]    4    1    2    2    1
#>  [12,]    4    1    2    1    2
#>  [13,]    4    1    1    3    1
#>  [14,]    4    1    1    2    2
#>  [15,]    4    1    1    1    3
#>  [16,]    3    4    1    1    1
#>  [17,]    3    3    2    1    1
#>  [18,]    3    3    1    2    1
#>  [19,]    3    3    1    1    2
#>  [20,]    3    2    3    1    1
#>  [21,]    3    2    2    2    1
#>  [22,]    3    2    2    1    2
#>  [23,]    3    2    1    3    1
#>  [24,]    3    2    1    2    2
#>  [25,]    3    2    1    1    3
#>  [26,]    3    1    4    1    1
#>  [27,]    3    1    3    2    1
#>  [28,]    3    1    3    1    2
#>  [29,]    3    1    2    3    1
#>  [30,]    3    1    2    2    2
#>  [31,]    3    1    2    1    3
#>  [32,]    3    1    1    4    1
#>  [33,]    3    1    1    3    2
#>  [34,]    3    1    1    2    3
#>  [35,]    3    1    1    1    4
#>  [36,]    2    5    1    1    1
#>  [37,]    2    4    2    1    1
#>  [38,]    2    4    1    2    1
#>  [39,]    2    4    1    1    2
#>  [40,]    2    3    3    1    1
#>  [41,]    2    3    2    2    1
#>  [42,]    2    3    2    1    2
#>  [43,]    2    3    1    3    1
#>  [44,]    2    3    1    2    2
#>  [45,]    2    3    1    1    3
#>  [46,]    2    2    4    1    1
#>  [47,]    2    2    3    2    1
#>  [48,]    2    2    3    1    2
#>  [49,]    2    2    2    3    1
#>  [50,]    2    2    2    2    2
#>  [51,]    2    2    2    1    3
#>  [52,]    2    2    1    4    1
#>  [53,]    2    2    1    3    2
#>  [54,]    2    2    1    2    3
#>  [55,]    2    2    1    1    4
#>  [56,]    2    1    5    1    1
#>  [57,]    2    1    4    2    1
#>  [58,]    2    1    4    1    2
#>  [59,]    2    1    3    3    1
#>  [60,]    2    1    3    2    2
#>  [61,]    2    1    3    1    3
#>  [62,]    2    1    2    4    1
#>  [63,]    2    1    2    3    2
#>  [64,]    2    1    2    2    3
#>  [65,]    2    1    2    1    4
#>  [66,]    2    1    1    5    1
#>  [67,]    2    1    1    4    2
#>  [68,]    2    1    1    3    3
#>  [69,]    2    1    1    2    4
#>  [70,]    2    1    1    1    5
#>  [71,]    1    6    1    1    1
#>  [72,]    1    5    2    1    1
#>  [73,]    1    5    1    2    1
#>  [74,]    1    5    1    1    2
#>  [75,]    1    4    3    1    1
#>  [76,]    1    4    2    2    1
#>  [77,]    1    4    2    1    2
#>  [78,]    1    4    1    3    1
#>  [79,]    1    4    1    2    2
#>  [80,]    1    4    1    1    3
#>  [81,]    1    3    4    1    1
#>  [82,]    1    3    3    2    1
#>  [83,]    1    3    3    1    2
#>  [84,]    1    3    2    3    1
#>  [85,]    1    3    2    2    2
#>  [86,]    1    3    2    1    3
#>  [87,]    1    3    1    4    1
#>  [88,]    1    3    1    3    2
#>  [89,]    1    3    1    2    3
#>  [90,]    1    3    1    1    4
#>  [91,]    1    2    5    1    1
#>  [92,]    1    2    4    2    1
#>  [93,]    1    2    4    1    2
#>  [94,]    1    2    3    3    1
#>  [95,]    1    2    3    2    2
#>  [96,]    1    2    3    1    3
#>  [97,]    1    2    2    4    1
#>  [98,]    1    2    2    3    2
#>  [99,]    1    2    2    2    3
#> [100,]    1    2    2    1    4
#> [101,]    1    2    1    5    1
#> [102,]    1    2    1    4    2
#> [103,]    1    2    1    3    3
#> [104,]    1    2    1    2    4
#> [105,]    1    2    1    1    5
#> [106,]    1    1    6    1    1
#> [107,]    1    1    5    2    1
#> [108,]    1    1    5    1    2
#> [109,]    1    1    4    3    1
#> [110,]    1    1    4    2    2
#> [111,]    1    1    4    1    3
#> [112,]    1    1    3    4    1
#> [113,]    1    1    3    3    2
#> [114,]    1    1    3    2    3
#> [115,]    1    1    3    1    4
#> [116,]    1    1    2    5    1
#> [117,]    1    1    2    4    2
#> [118,]    1    1    2    3    3
#> [119,]    1    1    2    2    4
#> [120,]    1    1    2    1    5
#> [121,]    1    1    1    6    1
#> [122,]    1    1    1    5    2
#> [123,]    1    1    1    4    3
#> [124,]    1    1    1    3    4
#> [125,]    1    1    1    2    5
#> [126,]    1    1    1    1    6

# column major
compositions(6, layout = "column")

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    1    1    1    1    1    1    1    1    1     1     1     1     1     1
#> [2,]    1    1    1    1    1    1    1    1    2     2     2     2     3     3
#> [3,]    1    1    1    1    2    2    3    4    1     1     2     3     1     2
#> [4,]    1    1    2    3    1    2    1    0    1     2     1     0     1     0
#> [5,]    1    2    1    0    1    0    0    0    1     0     0     0     0     0
#> [6,]    1    0    0    0    0    0    0    0    0     0     0     0     0     0
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]     1     1     2     2     2     2     2     2     2     2     3     3
#> [2,]     4     5     1     1     1     1     2     2     3     4     1     1
#> [3,]     1     0     1     1     2     3     1     2     1     0     1     2
#> [4,]     0     0     1     2     1     0     1     0     0     0     1     0
#> [5,]     0     0     1     0     0     0     0     0     0     0     0     0
#> [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,27] [,28] [,29] [,30] [,31] [,32]
#> [1,]     3     3     4     4     5     6
#> [2,]     2     3     1     2     1     0
#> [3,]     1     0     1     0     0     0
#> [4,]     0     0     0     0     0     0
#> [5,]     0     0     0     0     0     0
#> [6,]     0     0     0     0     0     0

compositions(6, 3, layout = "column")

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    1    1    1    1    2    2    2    3    3     4
#> [2,]    1    2    3    4    1    2    3    1    2     1
#> [3,]    4    3    2    1    3    2    1    2    1     1

# list output
compositions(6, layout = "list")

#> [[1]]
#> [1] 1 1 1 1 1 1
#> 
#> [[2]]
#> [1] 1 1 1 1 2
#> 
#> [[3]]
#> [1] 1 1 1 2 1
#> 
#> [[4]]
#> [1] 1 1 1 3
#> 
#> [[5]]
#> [1] 1 1 2 1 1
#> 
#> [[6]]
#> [1] 1 1 2 2
#> 
#> [[7]]
#> [1] 1 1 3 1
#> 
#> [[8]]
#> [1] 1 1 4
#> 
#> [[9]]
#> [1] 1 2 1 1 1
#> 
#> [[10]]
#> [1] 1 2 1 2
#> 
#> [[11]]
#> [1] 1 2 2 1
#> 
#> [[12]]
#> [1] 1 2 3
#> 
#> [[13]]
#> [1] 1 3 1 1
#> 
#> [[14]]
#> [1] 1 3 2
#> 
#> [[15]]
#> [1] 1 4 1
#> 
#> [[16]]
#> [1] 1 5
#> 
#> [[17]]
#> [1] 2 1 1 1 1
#> 
#> [[18]]
#> [1] 2 1 1 2
#> 
#> [[19]]
#> [1] 2 1 2 1
#> 
#> [[20]]
#> [1] 2 1 3
#> 
#> [[21]]
#> [1] 2 2 1 1
#> 
#> [[22]]
#> [1] 2 2 2
#> 
#> [[23]]
#> [1] 2 3 1
#> 
#> [[24]]
#> [1] 2 4
#> 
#> [[25]]
#> [1] 3 1 1 1
#> 
#> [[26]]
#> [1] 3 1 2
#> 
#> [[27]]
#> [1] 3 2 1
#> 
#> [[28]]
#> [1] 3 3
#> 
#> [[29]]
#> [1] 4 1 1
#> 
#> [[30]]
#> [1] 4 2
#> 
#> [[31]]
#> [1] 5 1
#> 
#> [[32]]
#> [1] 6
#> 

compositions(6, 3, layout = "list")

#> [[1]]
#> [1] 1 1 4
#> 
#> [[2]]
#> [1] 1 2 3
#> 
#> [[3]]
#> [1] 1 3 2
#> 
#> [[4]]
#> [1] 1 4 1
#> 
#> [[5]]
#> [1] 2 1 3
#> 
#> [[6]]
#> [1] 2 2 2
#> 
#> [[7]]
#> [1] 2 3 1
#> 
#> [[8]]
#> [1] 3 1 2
#> 
#> [[9]]
#> [1] 3 2 1
#> 
#> [[10]]
#> [1] 4 1 1
#> 

# zero sized compositions
dim(compositions(0))

#> [1] 1 0

dim(compositions(5, 0))

#> [1] 0 0

dim(compositions(5, 6))

#> [1] 0 6

dim(compositions(0, 0))

#> [1] 1 0

dim(compositions(0, 1))

#> [1] 0 1