Combinations iterator

This function returns a Combinations iterator which allows users to fetch the next combination(s) via the getnext() method. All remaing combinations of the iterator can be fetched via the collect() method.

Combinations

icombinations(n, k, x = NULL, f = NULL, replace = FALSE)

Arguments

n
an integer, would be determined implicitly from x or f if missing

k
an integer

x
an optional vector indicating item labels

f
an integer vector of item repeat frequencies

replace
an logical to draw items with replacement

Format

An object of class R6ClassGenerator of length 24.

Details

The Combinations class can be initialized by using the convenient wrapper icombinations or

Combinations$new(n, k, x = NULL, f = NULL, replace = FALSE)

getnext(d = 1L, type = NULL, drop = d == 1L && is.null(type))
collect(type = "r")
reset()

d

number of fetched arrangements

type

if "r", "c" or "l" is specified, the returned value would be a "row-major" matrix, a "column-major" matrix or a list respectively

drop

vectorize a matrix or unlist a list

See also

combinations for generating all combinations and ncombinations to calculate number of combinations

Examples

icomb <- icombinations(5, 2)
icomb$getnext()

#> [1] 1 2

icomb$getnext(2)

#>      [,1] [,2]
#> [1,]    1    3
#> [2,]    1    4

icomb$getnext(type = "c", drop = FALSE)

#>      [,1]
#> [1,]    1
#> [2,]    5

# collect remaining combinations
icomb$collect()

#>      [,1] [,2]
#> [1,]    2    3
#> [2,]    2    4
#> [3,]    2    5
#> [4,]    3    4
#> [5,]    3    5
#> [6,]    4    5

library(foreach)
foreach(x = icombinations(5, 2), .combine=c) %do% {
  sum(x)
}

#>  [1] 3 4 5 6 5 6 7 7 8 9