Compositions iterator

This function returns a Compositions iterator for iterating compositions of an non-negative integer n into k parts or parts of any sizes. The iterator allows users to fetch the next partition(s) via the getnext() method.

Compositions

icompositions(n, k = NULL, descending = FALSE, skip = NULL)

Arguments

n
an non-negative integer to be partitioned

k
number of parts

descending
logical to use reversed lexicographical order

skip
the number of compositions skipped

Format

An object of class R6ClassGenerator of length 25.

Details

The Compositions class can be initialized by using the convenient wrapper icompositions or

Compositions$new(n, k = NULL, descending = FALSE)

getnext(d = 1L, layout = NULL, drop = NULL)
collect(layout = "row")
reset()

d

number of fetched arrangements

layout

if "row", "column" or "list" is specified, the returned value would be a "row-major" matrix, a "column-major" matrix or a list respectively

drop

vectorize a matrix or unlist a list

See also

Examples

ipart <- icompositions(10)
ipart$getnext()

#>  [1] 1 1 1 1 1 1 1 1 1 1

ipart$getnext(2)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    1    1    1    1    1    1    1    1    2     0
#> [2,]    1    1    1    1    1    1    1    2    1     0

ipart$getnext(layout = "column", drop = FALSE)

#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    3
#>  [9,]    0
#> [10,]    0

# collect remaining compositions
ipart$collect()

#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>   [1,]    1    1    1    1    1    1    2    1    1     0
#>   [2,]    1    1    1    1    1    1    2    2    0     0
#>   [3,]    1    1    1    1    1    1    3    1    0     0
#>   [4,]    1    1    1    1    1    1    4    0    0     0
#>   [5,]    1    1    1    1    1    2    1    1    1     0
#>   [6,]    1    1    1    1    1    2    1    2    0     0
#>   [7,]    1    1    1    1    1    2    2    1    0     0
#>   [8,]    1    1    1    1    1    2    3    0    0     0
#>   [9,]    1    1    1    1    1    3    1    1    0     0
#>  [10,]    1    1    1    1    1    3    2    0    0     0
#>  [11,]    1    1    1    1    1    4    1    0    0     0
#>  [12,]    1    1    1    1    1    5    0    0    0     0
#>  [13,]    1    1    1    1    2    1    1    1    1     0
#>  [14,]    1    1    1    1    2    1    1    2    0     0
#>  [15,]    1    1    1    1    2    1    2    1    0     0
#>  [16,]    1    1    1    1    2    1    3    0    0     0
#>  [17,]    1    1    1    1    2    2    1    1    0     0
#>  [18,]    1    1    1    1    2    2    2    0    0     0
#>  [19,]    1    1    1    1    2    3    1    0    0     0
#>  [20,]    1    1    1    1    2    4    0    0    0     0
#>  [21,]    1    1    1    1    3    1    1    1    0     0
#>  [22,]    1    1    1    1    3    1    2    0    0     0
#>  [23,]    1    1    1    1    3    2    1    0    0     0
#>  [24,]    1    1    1    1    3    3    0    0    0     0
#>  [25,]    1    1    1    1    4    1    1    0    0     0
#>  [26,]    1    1    1    1    4    2    0    0    0     0
#>  [27,]    1    1    1    1    5    1    0    0    0     0
#>  [28,]    1    1    1    1    6    0    0    0    0     0
#>  [29,]    1    1    1    2    1    1    1    1    1     0
#>  [30,]    1    1    1    2    1    1    1    2    0     0
#>  [31,]    1    1    1    2    1    1    2    1    0     0
#>  [32,]    1    1    1    2    1    1    3    0    0     0
#>  [33,]    1    1    1    2    1    2    1    1    0     0
#>  [34,]    1    1    1    2    1    2    2    0    0     0
#>  [35,]    1    1    1    2    1    3    1    0    0     0
#>  [36,]    1    1    1    2    1    4    0    0    0     0
#>  [37,]    1    1    1    2    2    1    1    1    0     0
#>  [38,]    1    1    1    2    2    1    2    0    0     0
#>  [39,]    1    1    1    2    2    2    1    0    0     0
#>  [40,]    1    1    1    2    2    3    0    0    0     0
#>  [41,]    1    1    1    2    3    1    1    0    0     0
#>  [42,]    1    1    1    2    3    2    0    0    0     0
#>  [43,]    1    1    1    2    4    1    0    0    0     0
#>  [44,]    1    1    1    2    5    0    0    0    0     0
#>  [45,]    1    1    1    3    1    1    1    1    0     0
#>  [46,]    1    1    1    3    1    1    2    0    0     0
#>  [47,]    1    1    1    3    1    2    1    0    0     0
#>  [48,]    1    1    1    3    1    3    0    0    0     0
#>  [49,]    1    1    1    3    2    1    1    0    0     0
#>  [50,]    1    1    1    3    2    2    0    0    0     0
#>  [51,]    1    1    1    3    3    1    0    0    0     0
#>  [52,]    1    1    1    3    4    0    0    0    0     0
#>  [53,]    1    1    1    4    1    1    1    0    0     0
#>  [54,]    1    1    1    4    1    2    0    0    0     0
#>  [55,]    1    1    1    4    2    1    0    0    0     0
#>  [56,]    1    1    1    4    3    0    0    0    0     0
#>  [57,]    1    1    1    5    1    1    0    0    0     0
#>  [58,]    1    1    1    5    2    0    0    0    0     0
#>  [59,]    1    1    1    6    1    0    0    0    0     0
#>  [60,]    1    1    1    7    0    0    0    0    0     0
#>  [61,]    1    1    2    1    1    1    1    1    1     0
#>  [62,]    1    1    2    1    1    1    1    2    0     0
#>  [63,]    1    1    2    1    1    1    2    1    0     0
#>  [64,]    1    1    2    1    1    1    3    0    0     0
#>  [65,]    1    1    2    1    1    2    1    1    0     0
#>  [66,]    1    1    2    1    1    2    2    0    0     0
#>  [67,]    1    1    2    1    1    3    1    0    0     0
#>  [68,]    1    1    2    1    1    4    0    0    0     0
#>  [69,]    1    1    2    1    2    1    1    1    0     0
#>  [70,]    1    1    2    1    2    1    2    0    0     0
#>  [71,]    1    1    2    1    2    2    1    0    0     0
#>  [72,]    1    1    2    1    2    3    0    0    0     0
#>  [73,]    1    1    2    1    3    1    1    0    0     0
#>  [74,]    1    1    2    1    3    2    0    0    0     0
#>  [75,]    1    1    2    1    4    1    0    0    0     0
#>  [76,]    1    1    2    1    5    0    0    0    0     0
#>  [77,]    1    1    2    2    1    1    1    1    0     0
#>  [78,]    1    1    2    2    1    1    2    0    0     0
#>  [79,]    1    1    2    2    1    2    1    0    0     0
#>  [80,]    1    1    2    2    1    3    0    0    0     0
#>  [81,]    1    1    2    2    2    1    1    0    0     0
#>  [82,]    1    1    2    2    2    2    0    0    0     0
#>  [83,]    1    1    2    2    3    1    0    0    0     0
#>  [84,]    1    1    2    2    4    0    0    0    0     0
#>  [85,]    1    1    2    3    1    1    1    0    0     0
#>  [86,]    1    1    2    3    1    2    0    0    0     0
#>  [87,]    1    1    2    3    2    1    0    0    0     0
#>  [88,]    1    1    2    3    3    0    0    0    0     0
#>  [89,]    1    1    2    4    1    1    0    0    0     0
#>  [90,]    1    1    2    4    2    0    0    0    0     0
#>  [91,]    1    1    2    5    1    0    0    0    0     0
#>  [92,]    1    1    2    6    0    0    0    0    0     0
#>  [93,]    1    1    3    1    1    1    1    1    0     0
#>  [94,]    1    1    3    1    1    1    2    0    0     0
#>  [95,]    1    1    3    1    1    2    1    0    0     0
#>  [96,]    1    1    3    1    1    3    0    0    0     0
#>  [97,]    1    1    3    1    2    1    1    0    0     0
#>  [98,]    1    1    3    1    2    2    0    0    0     0
#>  [99,]    1    1    3    1    3    1    0    0    0     0
#> [100,]    1    1    3    1    4    0    0    0    0     0
#> [101,]    1    1    3    2    1    1    1    0    0     0
#> [102,]    1    1    3    2    1    2    0    0    0     0
#> [103,]    1    1    3    2    2    1    0    0    0     0
#> [104,]    1    1    3    2    3    0    0    0    0     0
#> [105,]    1    1    3    3    1    1    0    0    0     0
#> [106,]    1    1    3    3    2    0    0    0    0     0
#> [107,]    1    1    3    4    1    0    0    0    0     0
#> [108,]    1    1    3    5    0    0    0    0    0     0
#> [109,]    1    1    4    1    1    1    1    0    0     0
#> [110,]    1    1    4    1    1    2    0    0    0     0
#> [111,]    1    1    4    1    2    1    0    0    0     0
#> [112,]    1    1    4    1    3    0    0    0    0     0
#> [113,]    1    1    4    2    1    1    0    0    0     0
#> [114,]    1    1    4    2    2    0    0    0    0     0
#> [115,]    1    1    4    3    1    0    0    0    0     0
#> [116,]    1    1    4    4    0    0    0    0    0     0
#> [117,]    1    1    5    1    1    1    0    0    0     0
#> [118,]    1    1    5    1    2    0    0    0    0     0
#> [119,]    1    1    5    2    1    0    0    0    0     0
#> [120,]    1    1    5    3    0    0    0    0    0     0
#> [121,]    1    1    6    1    1    0    0    0    0     0
#> [122,]    1    1    6    2    0    0    0    0    0     0
#> [123,]    1    1    7    1    0    0    0    0    0     0
#> [124,]    1    1    8    0    0    0    0    0    0     0
#> [125,]    1    2    1    1    1    1    1    1    1     0
#> [126,]    1    2    1    1    1    1    1    2    0     0
#> [127,]    1    2    1    1    1    1    2    1    0     0
#> [128,]    1    2    1    1    1    1    3    0    0     0
#> [129,]    1    2    1    1    1    2    1    1    0     0
#> [130,]    1    2    1    1    1    2    2    0    0     0
#> [131,]    1    2    1    1    1    3    1    0    0     0
#> [132,]    1    2    1    1    1    4    0    0    0     0
#> [133,]    1    2    1    1    2    1    1    1    0     0
#> [134,]    1    2    1    1    2    1    2    0    0     0
#> [135,]    1    2    1    1    2    2    1    0    0     0
#> [136,]    1    2    1    1    2    3    0    0    0     0
#> [137,]    1    2    1    1    3    1    1    0    0     0
#> [138,]    1    2    1    1    3    2    0    0    0     0
#> [139,]    1    2    1    1    4    1    0    0    0     0
#> [140,]    1    2    1    1    5    0    0    0    0     0
#> [141,]    1    2    1    2    1    1    1    1    0     0
#> [142,]    1    2    1    2    1    1    2    0    0     0
#> [143,]    1    2    1    2    1    2    1    0    0     0
#> [144,]    1    2    1    2    1    3    0    0    0     0
#> [145,]    1    2    1    2    2    1    1    0    0     0
#> [146,]    1    2    1    2    2    2    0    0    0     0
#> [147,]    1    2    1    2    3    1    0    0    0     0
#> [148,]    1    2    1    2    4    0    0    0    0     0
#> [149,]    1    2    1    3    1    1    1    0    0     0
#> [150,]    1    2    1    3    1    2    0    0    0     0
#> [151,]    1    2    1    3    2    1    0    0    0     0
#> [152,]    1    2    1    3    3    0    0    0    0     0
#> [153,]    1    2    1    4    1    1    0    0    0     0
#> [154,]    1    2    1    4    2    0    0    0    0     0
#> [155,]    1    2    1    5    1    0    0    0    0     0
#> [156,]    1    2    1    6    0    0    0    0    0     0
#> [157,]    1    2    2    1    1    1    1    1    0     0
#> [158,]    1    2    2    1    1    1    2    0    0     0
#> [159,]    1    2    2    1    1    2    1    0    0     0
#> [160,]    1    2    2    1    1    3    0    0    0     0
#> [161,]    1    2    2    1    2    1    1    0    0     0
#> [162,]    1    2    2    1    2    2    0    0    0     0
#> [163,]    1    2    2    1    3    1    0    0    0     0
#> [164,]    1    2    2    1    4    0    0    0    0     0
#> [165,]    1    2    2    2    1    1    1    0    0     0
#> [166,]    1    2    2    2    1    2    0    0    0     0
#> [167,]    1    2    2    2    2    1    0    0    0     0
#> [168,]    1    2    2    2    3    0    0    0    0     0
#> [169,]    1    2    2    3    1    1    0    0    0     0
#> [170,]    1    2    2    3    2    0    0    0    0     0
#> [171,]    1    2    2    4    1    0    0    0    0     0
#> [172,]    1    2    2    5    0    0    0    0    0     0
#> [173,]    1    2    3    1    1    1    1    0    0     0
#> [174,]    1    2    3    1    1    2    0    0    0     0
#> [175,]    1    2    3    1    2    1    0    0    0     0
#> [176,]    1    2    3    1    3    0    0    0    0     0
#> [177,]    1    2    3    2    1    1    0    0    0     0
#> [178,]    1    2    3    2    2    0    0    0    0     0
#> [179,]    1    2    3    3    1    0    0    0    0     0
#> [180,]    1    2    3    4    0    0    0    0    0     0
#> [181,]    1    2    4    1    1    1    0    0    0     0
#> [182,]    1    2    4    1    2    0    0    0    0     0
#> [183,]    1    2    4    2    1    0    0    0    0     0
#> [184,]    1    2    4    3    0    0    0    0    0     0
#> [185,]    1    2    5    1    1    0    0    0    0     0
#> [186,]    1    2    5    2    0    0    0    0    0     0
#> [187,]    1    2    6    1    0    0    0    0    0     0
#> [188,]    1    2    7    0    0    0    0    0    0     0
#> [189,]    1    3    1    1    1    1    1    1    0     0
#> [190,]    1    3    1    1    1    1    2    0    0     0
#> [191,]    1    3    1    1    1    2    1    0    0     0
#> [192,]    1    3    1    1    1    3    0    0    0     0
#> [193,]    1    3    1    1    2    1    1    0    0     0
#> [194,]    1    3    1    1    2    2    0    0    0     0
#> [195,]    1    3    1    1    3    1    0    0    0     0
#> [196,]    1    3    1    1    4    0    0    0    0     0
#> [197,]    1    3    1    2    1    1    1    0    0     0
#> [198,]    1    3    1    2    1    2    0    0    0     0
#> [199,]    1    3    1    2    2    1    0    0    0     0
#> [200,]    1    3    1    2    3    0    0    0    0     0
#> [201,]    1    3    1    3    1    1    0    0    0     0
#> [202,]    1    3    1    3    2    0    0    0    0     0
#> [203,]    1    3    1    4    1    0    0    0    0     0
#> [204,]    1    3    1    5    0    0    0    0    0     0
#> [205,]    1    3    2    1    1    1    1    0    0     0
#> [206,]    1    3    2    1    1    2    0    0    0     0
#> [207,]    1    3    2    1    2    1    0    0    0     0
#> [208,]    1    3    2    1    3    0    0    0    0     0
#> [209,]    1    3    2    2    1    1    0    0    0     0
#> [210,]    1    3    2    2    2    0    0    0    0     0
#> [211,]    1    3    2    3    1    0    0    0    0     0
#> [212,]    1    3    2    4    0    0    0    0    0     0
#> [213,]    1    3    3    1    1    1    0    0    0     0
#> [214,]    1    3    3    1    2    0    0    0    0     0
#> [215,]    1    3    3    2    1    0    0    0    0     0
#> [216,]    1    3    3    3    0    0    0    0    0     0
#> [217,]    1    3    4    1    1    0    0    0    0     0
#> [218,]    1    3    4    2    0    0    0    0    0     0
#> [219,]    1    3    5    1    0    0    0    0    0     0
#> [220,]    1    3    6    0    0    0    0    0    0     0
#> [221,]    1    4    1    1    1    1    1    0    0     0
#> [222,]    1    4    1    1    1    2    0    0    0     0
#> [223,]    1    4    1    1    2    1    0    0    0     0
#> [224,]    1    4    1    1    3    0    0    0    0     0
#> [225,]    1    4    1    2    1    1    0    0    0     0
#> [226,]    1    4    1    2    2    0    0    0    0     0
#> [227,]    1    4    1    3    1    0    0    0    0     0
#> [228,]    1    4    1    4    0    0    0    0    0     0
#> [229,]    1    4    2    1    1    1    0    0    0     0
#> [230,]    1    4    2    1    2    0    0    0    0     0
#> [231,]    1    4    2    2    1    0    0    0    0     0
#> [232,]    1    4    2    3    0    0    0    0    0     0
#> [233,]    1    4    3    1    1    0    0    0    0     0
#> [234,]    1    4    3    2    0    0    0    0    0     0
#> [235,]    1    4    4    1    0    0    0    0    0     0
#> [236,]    1    4    5    0    0    0    0    0    0     0
#> [237,]    1    5    1    1    1    1    0    0    0     0
#> [238,]    1    5    1    1    2    0    0    0    0     0
#> [239,]    1    5    1    2    1    0    0    0    0     0
#> [240,]    1    5    1    3    0    0    0    0    0     0
#> [241,]    1    5    2    1    1    0    0    0    0     0
#> [242,]    1    5    2    2    0    0    0    0    0     0
#> [243,]    1    5    3    1    0    0    0    0    0     0
#> [244,]    1    5    4    0    0    0    0    0    0     0
#> [245,]    1    6    1    1    1    0    0    0    0     0
#> [246,]    1    6    1    2    0    0    0    0    0     0
#> [247,]    1    6    2    1    0    0    0    0    0     0
#> [248,]    1    6    3    0    0    0    0    0    0     0
#> [249,]    1    7    1    1    0    0    0    0    0     0
#> [250,]    1    7    2    0    0    0    0    0    0     0
#> [251,]    1    8    1    0    0    0    0    0    0     0
#> [252,]    1    9    0    0    0    0    0    0    0     0
#> [253,]    2    1    1    1    1    1    1    1    1     0
#> [254,]    2    1    1    1    1    1    1    2    0     0
#> [255,]    2    1    1    1    1    1    2    1    0     0
#> [256,]    2    1    1    1    1    1    3    0    0     0
#> [257,]    2    1    1    1    1    2    1    1    0     0
#> [258,]    2    1    1    1    1    2    2    0    0     0
#> [259,]    2    1    1    1    1    3    1    0    0     0
#> [260,]    2    1    1    1    1    4    0    0    0     0
#> [261,]    2    1    1    1    2    1    1    1    0     0
#> [262,]    2    1    1    1    2    1    2    0    0     0
#> [263,]    2    1    1    1    2    2    1    0    0     0
#> [264,]    2    1    1    1    2    3    0    0    0     0
#> [265,]    2    1    1    1    3    1    1    0    0     0
#> [266,]    2    1    1    1    3    2    0    0    0     0
#> [267,]    2    1    1    1    4    1    0    0    0     0
#> [268,]    2    1    1    1    5    0    0    0    0     0
#> [269,]    2    1    1    2    1    1    1    1    0     0
#> [270,]    2    1    1    2    1    1    2    0    0     0
#> [271,]    2    1    1    2    1    2    1    0    0     0
#> [272,]    2    1    1    2    1    3    0    0    0     0
#> [273,]    2    1    1    2    2    1    1    0    0     0
#> [274,]    2    1    1    2    2    2    0    0    0     0
#> [275,]    2    1    1    2    3    1    0    0    0     0
#> [276,]    2    1    1    2    4    0    0    0    0     0
#> [277,]    2    1    1    3    1    1    1    0    0     0
#> [278,]    2    1    1    3    1    2    0    0    0     0
#> [279,]    2    1    1    3    2    1    0    0    0     0
#> [280,]    2    1    1    3    3    0    0    0    0     0
#> [281,]    2    1    1    4    1    1    0    0    0     0
#> [282,]    2    1    1    4    2    0    0    0    0     0
#> [283,]    2    1    1    5    1    0    0    0    0     0
#> [284,]    2    1    1    6    0    0    0    0    0     0
#> [285,]    2    1    2    1    1    1    1    1    0     0
#> [286,]    2    1    2    1    1    1    2    0    0     0
#> [287,]    2    1    2    1    1    2    1    0    0     0
#> [288,]    2    1    2    1    1    3    0    0    0     0
#> [289,]    2    1    2    1    2    1    1    0    0     0
#> [290,]    2    1    2    1    2    2    0    0    0     0
#> [291,]    2    1    2    1    3    1    0    0    0     0
#> [292,]    2    1    2    1    4    0    0    0    0     0
#> [293,]    2    1    2    2    1    1    1    0    0     0
#> [294,]    2    1    2    2    1    2    0    0    0     0
#> [295,]    2    1    2    2    2    1    0    0    0     0
#> [296,]    2    1    2    2    3    0    0    0    0     0
#> [297,]    2    1    2    3    1    1    0    0    0     0
#> [298,]    2    1    2    3    2    0    0    0    0     0
#> [299,]    2    1    2    4    1    0    0    0    0     0
#> [300,]    2    1    2    5    0    0    0    0    0     0
#> [301,]    2    1    3    1    1    1    1    0    0     0
#> [302,]    2    1    3    1    1    2    0    0    0     0
#> [303,]    2    1    3    1    2    1    0    0    0     0
#> [304,]    2    1    3    1    3    0    0    0    0     0
#> [305,]    2    1    3    2    1    1    0    0    0     0
#> [306,]    2    1    3    2    2    0    0    0    0     0
#> [307,]    2    1    3    3    1    0    0    0    0     0
#> [308,]    2    1    3    4    0    0    0    0    0     0
#> [309,]    2    1    4    1    1    1    0    0    0     0
#> [310,]    2    1    4    1    2    0    0    0    0     0
#> [311,]    2    1    4    2    1    0    0    0    0     0
#> [312,]    2    1    4    3    0    0    0    0    0     0
#> [313,]    2    1    5    1    1    0    0    0    0     0
#> [314,]    2    1    5    2    0    0    0    0    0     0
#> [315,]    2    1    6    1    0    0    0    0    0     0
#> [316,]    2    1    7    0    0    0    0    0    0     0
#> [317,]    2    2    1    1    1    1    1    1    0     0
#> [318,]    2    2    1    1    1    1    2    0    0     0
#> [319,]    2    2    1    1    1    2    1    0    0     0
#> [320,]    2    2    1    1    1    3    0    0    0     0
#> [321,]    2    2    1    1    2    1    1    0    0     0
#> [322,]    2    2    1    1    2    2    0    0    0     0
#> [323,]    2    2    1    1    3    1    0    0    0     0
#> [324,]    2    2    1    1    4    0    0    0    0     0
#> [325,]    2    2    1    2    1    1    1    0    0     0
#> [326,]    2    2    1    2    1    2    0    0    0     0
#> [327,]    2    2    1    2    2    1    0    0    0     0
#> [328,]    2    2    1    2    3    0    0    0    0     0
#> [329,]    2    2    1    3    1    1    0    0    0     0
#> [330,]    2    2    1    3    2    0    0    0    0     0
#> [331,]    2    2    1    4    1    0    0    0    0     0
#> [332,]    2    2    1    5    0    0    0    0    0     0
#> [333,]    2    2    2    1    1    1    1    0    0     0
#> [334,]    2    2    2    1    1    2    0    0    0     0
#> [335,]    2    2    2    1    2    1    0    0    0     0
#> [336,]    2    2    2    1    3    0    0    0    0     0
#> [337,]    2    2    2    2    1    1    0    0    0     0
#> [338,]    2    2    2    2    2    0    0    0    0     0
#> [339,]    2    2    2    3    1    0    0    0    0     0
#> [340,]    2    2    2    4    0    0    0    0    0     0
#> [341,]    2    2    3    1    1    1    0    0    0     0
#> [342,]    2    2    3    1    2    0    0    0    0     0
#> [343,]    2    2    3    2    1    0    0    0    0     0
#> [344,]    2    2    3    3    0    0    0    0    0     0
#> [345,]    2    2    4    1    1    0    0    0    0     0
#> [346,]    2    2    4    2    0    0    0    0    0     0
#> [347,]    2    2    5    1    0    0    0    0    0     0
#> [348,]    2    2    6    0    0    0    0    0    0     0
#> [349,]    2    3    1    1    1    1    1    0    0     0
#> [350,]    2    3    1    1    1    2    0    0    0     0
#> [351,]    2    3    1    1    2    1    0    0    0     0
#> [352,]    2    3    1    1    3    0    0    0    0     0
#> [353,]    2    3    1    2    1    1    0    0    0     0
#> [354,]    2    3    1    2    2    0    0    0    0     0
#> [355,]    2    3    1    3    1    0    0    0    0     0
#> [356,]    2    3    1    4    0    0    0    0    0     0
#> [357,]    2    3    2    1    1    1    0    0    0     0
#> [358,]    2    3    2    1    2    0    0    0    0     0
#> [359,]    2    3    2    2    1    0    0    0    0     0
#> [360,]    2    3    2    3    0    0    0    0    0     0
#> [361,]    2    3    3    1    1    0    0    0    0     0
#> [362,]    2    3    3    2    0    0    0    0    0     0
#> [363,]    2    3    4    1    0    0    0    0    0     0
#> [364,]    2    3    5    0    0    0    0    0    0     0
#> [365,]    2    4    1    1    1    1    0    0    0     0
#> [366,]    2    4    1    1    2    0    0    0    0     0
#> [367,]    2    4    1    2    1    0    0    0    0     0
#> [368,]    2    4    1    3    0    0    0    0    0     0
#> [369,]    2    4    2    1    1    0    0    0    0     0
#> [370,]    2    4    2    2    0    0    0    0    0     0
#> [371,]    2    4    3    1    0    0    0    0    0     0
#> [372,]    2    4    4    0    0    0    0    0    0     0
#> [373,]    2    5    1    1    1    0    0    0    0     0
#> [374,]    2    5    1    2    0    0    0    0    0     0
#> [375,]    2    5    2    1    0    0    0    0    0     0
#> [376,]    2    5    3    0    0    0    0    0    0     0
#> [377,]    2    6    1    1    0    0    0    0    0     0
#> [378,]    2    6    2    0    0    0    0    0    0     0
#> [379,]    2    7    1    0    0    0    0    0    0     0
#> [380,]    2    8    0    0    0    0    0    0    0     0
#> [381,]    3    1    1    1    1    1    1    1    0     0
#> [382,]    3    1    1    1    1    1    2    0    0     0
#> [383,]    3    1    1    1    1    2    1    0    0     0
#> [384,]    3    1    1    1    1    3    0    0    0     0
#> [385,]    3    1    1    1    2    1    1    0    0     0
#> [386,]    3    1    1    1    2    2    0    0    0     0
#> [387,]    3    1    1    1    3    1    0    0    0     0
#> [388,]    3    1    1    1    4    0    0    0    0     0
#> [389,]    3    1    1    2    1    1    1    0    0     0
#> [390,]    3    1    1    2    1    2    0    0    0     0
#> [391,]    3    1    1    2    2    1    0    0    0     0
#> [392,]    3    1    1    2    3    0    0    0    0     0
#> [393,]    3    1    1    3    1    1    0    0    0     0
#> [394,]    3    1    1    3    2    0    0    0    0     0
#> [395,]    3    1    1    4    1    0    0    0    0     0
#> [396,]    3    1    1    5    0    0    0    0    0     0
#> [397,]    3    1    2    1    1    1    1    0    0     0
#> [398,]    3    1    2    1    1    2    0    0    0     0
#> [399,]    3    1    2    1    2    1    0    0    0     0
#> [400,]    3    1    2    1    3    0    0    0    0     0
#> [401,]    3    1    2    2    1    1    0    0    0     0
#> [402,]    3    1    2    2    2    0    0    0    0     0
#> [403,]    3    1    2    3    1    0    0    0    0     0
#> [404,]    3    1    2    4    0    0    0    0    0     0
#> [405,]    3    1    3    1    1    1    0    0    0     0
#> [406,]    3    1    3    1    2    0    0    0    0     0
#> [407,]    3    1    3    2    1    0    0    0    0     0
#> [408,]    3    1    3    3    0    0    0    0    0     0
#> [409,]    3    1    4    1    1    0    0    0    0     0
#> [410,]    3    1    4    2    0    0    0    0    0     0
#> [411,]    3    1    5    1    0    0    0    0    0     0
#> [412,]    3    1    6    0    0    0    0    0    0     0
#> [413,]    3    2    1    1    1    1    1    0    0     0
#> [414,]    3    2    1    1    1    2    0    0    0     0
#> [415,]    3    2    1    1    2    1    0    0    0     0
#> [416,]    3    2    1    1    3    0    0    0    0     0
#> [417,]    3    2    1    2    1    1    0    0    0     0
#> [418,]    3    2    1    2    2    0    0    0    0     0
#> [419,]    3    2    1    3    1    0    0    0    0     0
#> [420,]    3    2    1    4    0    0    0    0    0     0
#> [421,]    3    2    2    1    1    1    0    0    0     0
#> [422,]    3    2    2    1    2    0    0    0    0     0
#> [423,]    3    2    2    2    1    0    0    0    0     0
#> [424,]    3    2    2    3    0    0    0    0    0     0
#> [425,]    3    2    3    1    1    0    0    0    0     0
#> [426,]    3    2    3    2    0    0    0    0    0     0
#> [427,]    3    2    4    1    0    0    0    0    0     0
#> [428,]    3    2    5    0    0    0    0    0    0     0
#> [429,]    3    3    1    1    1    1    0    0    0     0
#> [430,]    3    3    1    1    2    0    0    0    0     0
#> [431,]    3    3    1    2    1    0    0    0    0     0
#> [432,]    3    3    1    3    0    0    0    0    0     0
#> [433,]    3    3    2    1    1    0    0    0    0     0
#> [434,]    3    3    2    2    0    0    0    0    0     0
#> [435,]    3    3    3    1    0    0    0    0    0     0
#> [436,]    3    3    4    0    0    0    0    0    0     0
#> [437,]    3    4    1    1    1    0    0    0    0     0
#> [438,]    3    4    1    2    0    0    0    0    0     0
#> [439,]    3    4    2    1    0    0    0    0    0     0
#> [440,]    3    4    3    0    0    0    0    0    0     0
#> [441,]    3    5    1    1    0    0    0    0    0     0
#> [442,]    3    5    2    0    0    0    0    0    0     0
#> [443,]    3    6    1    0    0    0    0    0    0     0
#> [444,]    3    7    0    0    0    0    0    0    0     0
#> [445,]    4    1    1    1    1    1    1    0    0     0
#> [446,]    4    1    1    1    1    2    0    0    0     0
#> [447,]    4    1    1    1    2    1    0    0    0     0
#> [448,]    4    1    1    1    3    0    0    0    0     0
#> [449,]    4    1    1    2    1    1    0    0    0     0
#> [450,]    4    1    1    2    2    0    0    0    0     0
#> [451,]    4    1    1    3    1    0    0    0    0     0
#> [452,]    4    1    1    4    0    0    0    0    0     0
#> [453,]    4    1    2    1    1    1    0    0    0     0
#> [454,]    4    1    2    1    2    0    0    0    0     0
#> [455,]    4    1    2    2    1    0    0    0    0     0
#> [456,]    4    1    2    3    0    0    0    0    0     0
#> [457,]    4    1    3    1    1    0    0    0    0     0
#> [458,]    4    1    3    2    0    0    0    0    0     0
#> [459,]    4    1    4    1    0    0    0    0    0     0
#> [460,]    4    1    5    0    0    0    0    0    0     0
#> [461,]    4    2    1    1    1    1    0    0    0     0
#> [462,]    4    2    1    1    2    0    0    0    0     0
#> [463,]    4    2    1    2    1    0    0    0    0     0
#> [464,]    4    2    1    3    0    0    0    0    0     0
#> [465,]    4    2    2    1    1    0    0    0    0     0
#> [466,]    4    2    2    2    0    0    0    0    0     0
#> [467,]    4    2    3    1    0    0    0    0    0     0
#> [468,]    4    2    4    0    0    0    0    0    0     0
#> [469,]    4    3    1    1    1    0    0    0    0     0
#> [470,]    4    3    1    2    0    0    0    0    0     0
#> [471,]    4    3    2    1    0    0    0    0    0     0
#> [472,]    4    3    3    0    0    0    0    0    0     0
#> [473,]    4    4    1    1    0    0    0    0    0     0
#> [474,]    4    4    2    0    0    0    0    0    0     0
#> [475,]    4    5    1    0    0    0    0    0    0     0
#> [476,]    4    6    0    0    0    0    0    0    0     0
#> [477,]    5    1    1    1    1    1    0    0    0     0
#> [478,]    5    1    1    1    2    0    0    0    0     0
#> [479,]    5    1    1    2    1    0    0    0    0     0
#> [480,]    5    1    1    3    0    0    0    0    0     0
#> [481,]    5    1    2    1    1    0    0    0    0     0
#> [482,]    5    1    2    2    0    0    0    0    0     0
#> [483,]    5    1    3    1    0    0    0    0    0     0
#> [484,]    5    1    4    0    0    0    0    0    0     0
#> [485,]    5    2    1    1    1    0    0    0    0     0
#> [486,]    5    2    1    2    0    0    0    0    0     0
#> [487,]    5    2    2    1    0    0    0    0    0     0
#> [488,]    5    2    3    0    0    0    0    0    0     0
#> [489,]    5    3    1    1    0    0    0    0    0     0
#> [490,]    5    3    2    0    0    0    0    0    0     0
#> [491,]    5    4    1    0    0    0    0    0    0     0
#> [492,]    5    5    0    0    0    0    0    0    0     0
#> [493,]    6    1    1    1    1    0    0    0    0     0
#> [494,]    6    1    1    2    0    0    0    0    0     0
#> [495,]    6    1    2    1    0    0    0    0    0     0
#> [496,]    6    1    3    0    0    0    0    0    0     0
#> [497,]    6    2    1    1    0    0    0    0    0     0
#> [498,]    6    2    2    0    0    0    0    0    0     0
#> [499,]    6    3    1    0    0    0    0    0    0     0
#> [500,]    6    4    0    0    0    0    0    0    0     0
#> [501,]    7    1    1    1    0    0    0    0    0     0
#> [502,]    7    1    2    0    0    0    0    0    0     0
#> [503,]    7    2    1    0    0    0    0    0    0     0
#> [504,]    7    3    0    0    0    0    0    0    0     0
#> [505,]    8    1    1    0    0    0    0    0    0     0
#> [506,]    8    2    0    0    0    0    0    0    0     0
#> [507,]    9    1    0    0    0    0    0    0    0     0
#> [508,]   10    0    0    0    0    0    0    0    0     0

library(foreach)

#> Error in library(foreach): there is no package called ‘foreach’

foreach(x = icompositions(6, 2), .combine=c) %do% {
  prod(x)
}

#> Error in foreach(x = icompositions(6, 2), .combine = c) %do% {    prod(x)}: could not find function "%do%"