Permutations generator

This function generates all the permutations of selecting k items from n items. The results are in lexicographical order.

permutations(n, k = n, x = NULL, f = NULL, replace = FALSE,
  type = "r")

Arguments

n
an integer, would be determined implicitly from x or f if missing

k
an integer

x
an optional vector indicating item labels

f
an integer vector of item repeat frequencies

replace
an logical to draw items with replacement

type
if "r", "c" or "l" is specified, the returned value would be a "row-major" matrix, a "column-major" matrix or a list respectively

See also

ipermutations for iterating permutations and npermutations to calculate number of permutations

Examples

permutations(3)

#>      [,1] [,2] [,3]
#> [1,]    1    2    3
#> [2,]    1    3    2
#> [3,]    2    1    3
#> [4,]    2    3    1
#> [5,]    3    1    2
#> [6,]    3    2    1

permutations(x = LETTERS[1:3])

#>      [,1] [,2] [,3]
#> [1,] "A"  "B"  "C" 
#> [2,] "A"  "C"  "B" 
#> [3,] "B"  "A"  "C" 
#> [4,] "B"  "C"  "A" 
#> [5,] "C"  "A"  "B" 
#> [6,] "C"  "B"  "A" 

# choose 2 from 4
permutations(4, 2)

#>       [,1] [,2]
#>  [1,]    1    2
#>  [2,]    1    3
#>  [3,]    1    4
#>  [4,]    2    1
#>  [5,]    2    3
#>  [6,]    2    4
#>  [7,]    3    1
#>  [8,]    3    2
#>  [9,]    3    4
#> [10,]    4    1
#> [11,]    4    2
#> [12,]    4    3

permutations(x = LETTERS[1:3], k = 2)

#>      [,1] [,2]
#> [1,] "A"  "B" 
#> [2,] "A"  "C" 
#> [3,] "B"  "A" 
#> [4,] "B"  "C" 
#> [5,] "C"  "A" 
#> [6,] "C"  "B" 

# multiset with frequencies c(2, 3)
permutations(f = c(2, 3), k = 3)

#>      [,1] [,2] [,3]
#> [1,]    1    1    2
#> [2,]    1    2    1
#> [3,]    1    2    2
#> [4,]    2    1    1
#> [5,]    2    1    2
#> [6,]    2    2    1
#> [7,]    2    2    2

# with replacement
permutations(4, 2, replace = TRUE)

#>       [,1] [,2]
#>  [1,]    1    1
#>  [2,]    1    2
#>  [3,]    1    3
#>  [4,]    1    4
#>  [5,]    2    1
#>  [6,]    2    2
#>  [7,]    2    3
#>  [8,]    2    4
#>  [9,]    3    1
#> [10,]    3    2
#> [11,]    3    3
#> [12,]    3    4
#> [13,]    4    1
#> [14,]    4    2
#> [15,]    4    3
#> [16,]    4    4

# column major
permutations(3, type = "c")

#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1    1    2    2    3    3
#> [2,]    2    3    1    3    1    2
#> [3,]    3    2    3    1    2    1

permutations(4, 2, type = "c")

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
#> [1,]    1    1    1    2    2    2    3    3    3     4     4     4
#> [2,]    2    3    4    1    3    4    1    2    4     1     2     3

# list output
permutations(3, type = "l")

#> [[1]]
#> [1] 1 2 3
#> 
#> [[2]]
#> [1] 1 3 2
#> 
#> [[3]]
#> [1] 2 1 3
#> 
#> [[4]]
#> [1] 2 3 1
#> 
#> [[5]]
#> [1] 3 1 2
#> 
#> [[6]]
#> [1] 3 2 1
#> 

permutations(4, 2, type = "l")

#> [[1]]
#> [1] 1 2
#> 
#> [[2]]
#> [1] 1 3
#> 
#> [[3]]
#> [1] 1 4
#> 
#> [[4]]
#> [1] 2 1
#> 
#> [[5]]
#> [1] 2 3
#> 
#> [[6]]
#> [1] 2 4
#> 
#> [[7]]
#> [1] 3 1
#> 
#> [[8]]
#> [1] 3 2
#> 
#> [[9]]
#> [1] 3 4
#> 
#> [[10]]
#> [1] 4 1
#> 
#> [[11]]
#> [1] 4 2
#> 
#> [[12]]
#> [1] 4 3
#> 

# zero sized partitions
dim(permutations(0))

#> [1] 1 0

dim(permutations(5, 0))

#> [1] 1 0

dim(permutations(5, 6))

#> [1] 0 6

dim(permutations(0, 0))

#> [1] 1 0

dim(permutations(0, 1))

#> [1] 0 1